feat: expose validation error
