+++ lib/schemas/index.js
@@ -3,7 +3,6 @@
 const schema = require('./eikjson.schema.json');
 const validate = require('./validate');
 const assert = require('./assert');
+const ValidationError = require('./validation-error');
 
-module.exports.schema = schema;
-module.exports.validate = validate;
-module.exports.assert = assert;
+module.exports = { schema, validate, assert, ValidationError };
