+++ lib/schemas/validation-error.js
@@ -1,0 +1,12 @@
+'use strict';
+
+class ValidationError extends Error {
+    constructor(message, err) {
+        let m = message;
+        if (err && err.message) m += `: ${err.message}`;
+        super(m);
+        this.name = this.constructor.name;
+        Error.captureStackTrace(this, this.constructor);
+    }
+}
+module.exports = ValidationError;
